pyproject.toml
[tool.poetry]
name = "test-for-emphasoft"
version = "0.1.0"
description = ""
authors = ["flezzzer <141737861+flezzzer@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
psycopg2 = "^2.9.9"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-allauth = "^0.63.3"
django-filter = "^24.2"
pylint = "^3.2.5"
black = "^24.4.2"
mypy = "^1.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
